requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
colorama==0.4.6
discord==2.3.2
discord.py==2.6.4
filelock==3.20.0
frozenlist==1.8.0
google-auth==2.43.0
google-genai==1.50.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lmdb==1.7.5
multidict==6.7.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyfiglet==1.0.4
python-dotenv==1.2.1
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
shodan==1.31.0
sniffio==1.3.1
tenacity==9.1.2
tldextract==5.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.22.0
